Reorganising to support package format
